feat: 변수명을 헷갈리지 않게 돌려놓음
Done
- 컬럼 -> Num으로 통일
  - 학습 코드와 동일하게 가고, 추후 이름대로 쓸수있게 할 예정
