+++ src/main/kotlin/com/s3ich4n/theater/control/CheckAvailability.kt
@@ -1,14 +1,14 @@
 package com.s3ich4n.theater.control
 
-import TheaterData
+import com.s3ich4n.theater.domain.TheaterData
 
 class CheckAvailabilityBackingBean() {
     private val theaterData: TheaterData = TheaterData()
 
-    val seatCols = 1..theaterData.theaterColumns
+    val seatNums = 1..theaterData.theaterNums
     val seatRows = 'A'..'O'
 
-    var selectedSeatColumn : Int = 1
+    var selectedSeatNum : Int = 1
     var selectedSeatRow : Char = 'A'
     var result : String = ""
 }
