+++ src/main/kotlin/com/s3ich4n/theater/domain/Seat.kt
@@ -1,0 +1,7 @@
+package com.s3ich4n.theater.domain
+
+import java.math.BigDecimal
+
+data class Seat(val row: Char, val num: Int, val price: BigDecimal, val description: String) {
+    override fun toString(): String = "Seat $row-$num $$price ($description)"
+}
